feat: delete any poll when you're a moderator
This feature is to be discussed.
It would come on handy to deal with spam,
but it creates leeway for moderation abuses.
It can be mitigated by the circle
User --> Observatory --> Modo --> User --> …
Moderators would have to answer to an Observatory (sort of council?)
and Users may complain to the Observatory.